Make generate as a sub-command to prometheus
